More fine grained subscriptions and iron router waitOns
Previously the client subscribed to the entire Problems and Users collections,
which resulted in poor loads on crappy networks. Now the subscriptions are more
fine grained so the clients only fetch what they actually need.

The poor loading times for data also resulted in the client rendering the site
before the data was ready, leaving "holes" in the site. We now utilize the
waitOn hook in iron-router to rectify this.