Removed libphonenumber, as we need to validate on the client as well due to long polling instead of web sockets on the server
There is a bug on the prod server where throwing a meteor error doesn't seem to
return enough data for the useraccounts package to know in which field the
validation error occurred. Maybe it has something to do with long polling?

Anyway, since the phone number was only validated on the server side due to the
nature of libphonenumber, this meant that the for the user everything seemed
okay, but it didn't, and he wouldn't know which field didn't validate (if the
phone number was invalid).

Use simple regex check for now, it doesn't matter really